requirements.txt
# psycopg2-binary==2.8.6

django==2.2.*
Pillow==7.2.0
requests==2.20
GitPython==3.1.18
